Moved some internal forwardRef tests to not be internal
