+++ packages/react/src/__tests__/forwardRef-test.internal.js
@@ -94,31 +94,6 @@ describe('forwardRef', () => {
     expect(ref.current instanceof Child).toBe(true);
   });
 
-  it('should update refs when switching between children', () => {
-    function FunctionalComponent({forwardedRef, setRefOnDiv}) {
-      return (
-        <section>
-          <div ref={setRefOnDiv ? forwardedRef : null}>First</div>
-          <span ref={setRefOnDiv ? null : forwardedRef}>Second</span>
-        </section>
-      );
-    }
-
-    const RefForwardingComponent = React.forwardRef((props, ref) => (
-      <FunctionalComponent {...props} forwardedRef={ref} />
-    ));
-
-    const ref = React.createRef();
-
-    ReactNoop.render(<RefForwardingComponent ref={ref} setRefOnDiv={true} />);
-    ReactNoop.flush();
-    expect(ref.current.type).toBe('div');
-
-    ReactNoop.render(<RefForwardingComponent ref={ref} setRefOnDiv={false} />);
-    ReactNoop.flush();
-    expect(ref.current.type).toBe('span');
-  });
-
   it('should maintain child instance and ref through updates', () => {
     class Child extends React.Component {
       constructor(props) {
@@ -206,32 +181,6 @@ describe('forwardRef', () => {
     expect(ref.current).toBe(null);
   });
 
-  it('should support rendering null', () => {
-    const RefForwardingComponent = React.forwardRef((props, ref) => null);
-
-    const ref = React.createRef();
-
-    ReactNoop.render(<RefForwardingComponent ref={ref} />);
-    ReactNoop.flush();
-    expect(ref.current).toBe(null);
-  });
-
-  it('should support rendering null for multiple children', () => {
-    const RefForwardingComponent = React.forwardRef((props, ref) => null);
-
-    const ref = React.createRef();
-
-    ReactNoop.render(
-      <div>
-        <div />
-        <RefForwardingComponent ref={ref} />
-        <div />
-      </div>,
-    );
-    ReactNoop.flush();
-    expect(ref.current).toBe(null);
-  });
-
   it('should not re-run the render callback on a deep setState', () => {
     let inst;
 
@@ -263,44 +212,5 @@ describe('forwardRef', () => {
 
     inst.setState({});
     expect(ReactNoop.flush()).toEqual(['Inner']);
-  });
-
-  it('should warn if not provided a callback during creation', () => {
-    expect(() => React.forwardRef(undefined)).toWarnDev(
-      'forwardRef requires a render function but was given undefined.',
-    );
-    expect(() => React.forwardRef(null)).toWarnDev(
-      'forwardRef requires a render function but was given null.',
-    );
-    expect(() => React.forwardRef('foo')).toWarnDev(
-      'forwardRef requires a render function but was given string.',
-    );
-  });
-
-  it('should warn if no render function is provided', () => {
-    expect(React.forwardRef).toWarnDev(
-      'forwardRef requires a render function but was given undefined.',
-    );
-  });
-
-  it('should warn if the render function provided has propTypes or defaultProps attributes', () => {
-    function renderWithPropTypes() {
-      return null;
-    }
-    renderWithPropTypes.propTypes = {};
-
-    function renderWithDefaultProps() {
-      return null;
-    }
-    renderWithDefaultProps.defaultProps = {};
-
-    expect(() => React.forwardRef(renderWithPropTypes)).toWarnDev(
-      'forwardRef render functions do not support propTypes or defaultProps. ' +
-        'Did you accidentally pass a React component?',
-    );
-    expect(() => React.forwardRef(renderWithDefaultProps)).toWarnDev(
-      'forwardRef render functions do not support propTypes or defaultProps. ' +
-        'Did you accidentally pass a React component?',
-    );
   });
 });
